Updated REST Views routes
All /content/views routes are moved to /views.
BC is maintained for /content/views as it was already implemented.

Actions are moved to their own Controller.

EZP-24671
