+++ eZ/Bundle/EzPublishRestBundle/Resources/config/services.yml
@@ -20,6 +20,7 @@ parameters:
     ezpublish_rest.controller.url_alias.class: eZ\Publish\Core\REST\Server\Controller\URLAlias
     ezpublish_rest.controller.options.class: eZ\Publish\Core\REST\Server\Controller\Options
     ezpublish_rest.controller.services.class: eZ\Publish\Core\REST\Server\Controller\Services
+    ezpublish_rest.controller.views.class: eZ\Publish\Core\REST\Server\Controller\Views
 
     ezpublish_rest.factory.class: eZ\Bundle\EzPublishRestBundle\ApiLoader\Factory
     ezpublish_rest.request_parser.class: eZ\Bundle\EzPublishRestBundle\RequestParser\Router
@@ -209,6 +210,12 @@ services:
         arguments:
             - @ezpublish.api.service.url_alias
             - @ezpublish.api.service.location
+
+    ezpublish_rest.controller.views:
+        class: %ezpublish_rest.controller.views.class%
+        parent: ezpublish_rest.controller.base
+        arguments:
+            - @ezpublish.api.service.search
 
     ezpublish_rest.request_listener:
         class: %ezpublish_rest.request_listener.class%
